Fix bug in oligo padding function.
In the original release, the padding function failed to account for DNA sequences that formed when padding was joined with primer and oligo sequences. Very ocassionally, an extra restriction enzyme would be created.

I've updated the padding function to check for restriction sites in the
padding and a little bit where it overlaps into the primer and oligo
sequence. Chagnes originally made on /bug/padding branch and will be
merged later.